python/requirements.txt
ijson==2.3
typedb-client